requirements.txt
flask
subprocess
pyyaml
pyxdg
